[Fix #17] Don't add non-title entries to imenu
Fixes #17 by checking that 2-line titles have the same number of
under characters as the title text and doesn't look like the start
of an attribute block.

Also check when (adoc-title-descriptor) returns nil, which it will at
times when the regexes match but more stringent tests fail, a nil list
will be inserted into the imenu list, which it doesn't like much, so
this condition is checked for as well.